+++ stores/stores-service/stores-microservice-ejb/src/main/java/org/cocome/storesservice/repository/TradingEnterpriseDBRepository.java
@@ -46,6 +46,7 @@ public class TradingEnterpriseDBRepository implements TradingEnterpriseRepositor
 	public Long create(TradingEnterprise entity) {
 		try {
 			em.persist(entity);
+			em.flush();
 		} catch (Exception e) {
 			LOG.error("DATABASE: Database Error while creating Enterprise with name: " + entity.getName() + "     "
 					+ e.getMessage());
@@ -86,14 +87,18 @@ public class TradingEnterpriseDBRepository implements TradingEnterpriseRepositor
 	}
 
 	@Override
-	public void delete(Long key) {
+	public boolean delete(Long key) {
 		try {
 			TradingEnterprise entity = find(key);
-			em.remove(entity);
+			if (entity != null) {
+				em.remove(entity);
+				return true;
+			}
+
 		} catch (Exception e) {
 			LOG.error("DATABASE: Could not delete TradinEnterprise with id: " + key + "     " + e.getMessage());
 		}
-
+		return false;
 	}
 
 	@Override
